ADD InformationList페이 추가 + slide libary 사용

+++ src/components/ExerciseDiary/DiaryList.js
@@ -1,13 +1,27 @@
 import React, { useContext, useState } from "react";
 import DiaryContext from "../../store/diary-context";
-import Modal from "../Modal";
 import DiaryItem from "./DiaryItem";
-import EditDiary from "./EditDiary";
 
 const DiaryList = () => {
+  const todayDate = new Date();
+
   const ctx = useContext(DiaryContext);
   const handleDeleteButton = (e) => {
-    ctx.onDelete(e.target.id);
+    if (window.confirm("정말 제거하시겠습니까?")) {
+      ctx.onDelete(e.target.id);
+    }
+  };
+
+  const handleAddButton = () => {
+    ctx.addDiaryList({
+      date: `${todayDate.getFullYear()}-${(
+        "00" + String(todayDate.getMonth() + 1)
+      ).slice(-2)}-${("00" + todayDate.getDate()).slice(-2)}`,
+      part: "등",
+      minutes: "45",
+      satisfaction: "10",
+      review: "매우 만족",
+    });
   };
 
   return (
@@ -17,9 +31,14 @@ const DiaryList = () => {
           key={diary.id}
           handleDeleteButton={handleDeleteButton}
           diary={diary}
-          onEditDate={ctx.onEditDate}
+          onEdit={ctx.onEdit}
         />
       ))}
+      <div className="add--button-area">
+        <button className="add--button" onClick={handleAddButton}>
+          +
+        </button>
+      </div>
     </div>
   );
 };
